refactor: 移动 EconomySystem 文件
