+++ src/plotcraft/command/MergeCommand.cc
@@ -4,16 +4,17 @@
 #include "mc/world/actor/Actor.h"
 #include "mc/world/actor/player/Player.h"
 #include "plotcraft/Config.h"
+#include "plotcraft/EconomySystem.h"
 #include "plotcraft/Global.h"
 #include "plotcraft/data/PlotDBStorage.h"
 #include "plotcraft/math/PlotCross.h"
 #include "plotcraft/math/PlotPos.h"
 #include "plotcraft/math/PlotRoad.h"
-#include "plotcraft/utils/EconomySystem.h"
 #include "plotcraft/utils/Mc.h"
 #include <memory>
 #include <unordered_map>
 #include <utility>
+
 
 
 namespace plot::command {
