+++ src/plotcraft/gui/Global.h
@@ -10,6 +10,7 @@
 #include "mc/world/item/ItemStackBase.h"
 #include "mc/world/level/dimension/VanillaDimensions.h"
 #include "plotcraft/Config.h"
+#include "plotcraft/EconomySystem.h"
 #include "plotcraft/Global.h"
 #include "plotcraft/data/PlayerNameDB.h"
 #include "plotcraft/data/PlotDBStorage.h"
@@ -17,7 +18,6 @@
 #include "plotcraft/event/PlotEvents.h"
 #include "plotcraft/math/PlotPos.h"
 #include "plotcraft/utils/Date.h"
-#include "plotcraft/utils/EconomySystem.h"
 #include "plotcraft/utils/JsonHelper.h"
 #include "plotcraft/utils/Mc.h"
 #include "plotcraft/utils/Utils.h"
@@ -27,6 +27,7 @@
 #include <memory>
 #include <regex>
 #include <string>
+
 
 
 namespace plot::gui {
