+++ src/plugin/MyPlugin.cpp
@@ -18,14 +18,15 @@
 #include "mc/server/commands/ServerCommandOrigin.h"
 #include "mc/world/Minecraft.h"
 #include "plotcraft/Config.h"
+#include "plotcraft/EconomySystem.h"
 #include "plotcraft/command/Command.h"
 #include "plotcraft/core/PlotDimension.h"
 #include "plotcraft/data/PlayerNameDB.h"
 #include "plotcraft/data/PlotDBStorage.h"
 #include "plotcraft/event/Event.h"
-#include "plotcraft/utils/EconomySystem.h"
 #include "plotcraft/utils/Mc.h"
 #include <ll/api/utils/SystemUtils.h>
+
 
 #include "plotcraft/core/TemplateManager.h"
